requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]